+++ client/src/@types/Login.d.ts
@@ -1,0 +1,4 @@
+interface Login {
+  password: string;
+  email: string;
+}
