+++ client/src/@types/Profile.d.ts
@@ -1,0 +1,1 @@
+interface User {}
