+++ client/src/@types/Signup.d.ts
@@ -1,0 +1,9 @@
+interface Signup {
+  firstName: string;
+  lastName: string;
+  username: string;
+  password: string;
+  email: string;
+  image: string;
+  isAdmin: boolean;
+}
