+++ client/src/@types/index.d.ts
@@ -1,49 +1,46 @@
-enum Availability {
-  daily,
-  weekly,
-  biweekly,
-  monthly,
-  other,
+namespace Mentoring {
+  enum Availability {
+    daily = "daily",
+    weekly = "weekly",
+    biweekly = "biweekly",
+    monthly = "monthly",
+    other = "other",
+  }
+
+  type Level = "junior" | "intermediate" | "senior";
+
+  enum Field {
+    Frontend = "Frontend",
+    Backend = "Backend",
+    DevOps = "DevOps",
+    Softskills = "Softskills",
+  }
+
+  enum TechKnowHow {}
+
+  enum TimeSlots {}
+  interface Offer {
+    field: (keyof typeof Field)[];
+    techKnowHow: (keyof typeof TechKnowHow)[];
+    location: string;
+    yearsExp: number;
+    availability: (keyof typeof Availability)[];
+    timeslots: (keyof typeof TimeSlots)[];
+    level: Level;
+    description: string;
+    startDate: Date;
+  }
+  interface Request {
+    field: (keyof typeof Field)[];
+    techKnowHow: (keyof typeof TechKnowHow)[];
+    location: string;
+    yearsExp: number;
+    availability: (keyof typeof Availability)[];
+    timeslots: (keyof typeof TimeSlots)[];
+    description: string;
+    startDate: Date;
+  }
 }
 
-type Level = "junior" | "intermediate" | "senior";
-
-enum TechKnowHow {}
-
-enum TimeSlots {}
-
-interface SignUp {
-  firstName: string;
-  lastName: string;
-  username: string;
-  password: string;
-  email: string;
-  image: string;
-  isAdmin: boolean;
-}
-
-interface Login {
-  password: string;
-  email: string;
-}
-
-interface Offer {
-  techKnowHow: (keyof typeof TechKnowHow)[];
-  location: string;
-  yearsExp: number;
-  availability: (keyof typeof Availability)[];
-  timeslots: (keyof typeof TimeSlots)[];
-  level: Level;
-  description: string;
-  startDate: Date;
-}
-
-interface Request {
-  techKnowHow: (keyof typeof TechKnowHow)[];
-  location: string;
-  yearsExp: number;
-  availability: (keyof typeof Availability)[];
-  timeslots: (keyof typeof TimeSlots)[];
-  description: string;
-  startDate: Date;
-}
+namespace Coworking {}
+namespace Shadowing {}
