+++ client/src/Views/Profile.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Profile: React.FC = () => {
+  return <div>Signup</div>;
+};
+
+export default Profile;
